move ProductFieldsController to Platforms/OpenAI
